Also test Http2SolrClient async
