+++ solr/solrj/src/java/org/apache/solr/client/solrj/impl/HttpSolrClientBase.java
@@ -39,6 +39,8 @@ import org.apache.solr.client.solrj.SolrRequest;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.request.RequestWriter;
 import org.apache.solr.client.solrj.request.V2Request;
+import org.apache.solr.client.solrj.util.AsyncListener;
+import org.apache.solr.client.solrj.util.Cancellable;
 import org.apache.solr.common.SolrException;
 import org.apache.solr.common.params.CommonParams;
 import org.apache.solr.common.params.ModifiableSolrParams;
@@ -367,6 +369,19 @@ public abstract class HttpSolrClientBase extends SolrClient {
   }
 
   protected abstract void updateDefaultMimeTypeForParser();
+
+  /**
+   * Execute an asynchronous request to a Solr collection
+   *
+   * @param solrRequest the request to perform
+   * @param collection if null the default collection is used
+   * @param asyncListener callers should provide an implementation to handle events: start, success, exception
+   * @return Cancellable allowing the caller to attempt cancellation
+   */
+  public abstract Cancellable asyncRequest(
+          SolrRequest<?> solrRequest,
+          String collection,
+          AsyncListener<NamedList<Object>> asyncListener) ;
 
   public boolean isV2ApiRequest(final SolrRequest<?> request) {
     return request instanceof V2Request || request.getPath().contains("/____v2");
