+++ solr/solrj/src/test/org/apache/solr/client/solrj/impl/Http2SolrClientTest.java
@@ -264,6 +264,26 @@ public class Http2SolrClientTest extends HttpSolrClientTestBase {
   }
 
   @Test
+  public void testAsyncGet() throws Exception {
+    super.testQueryAsync(SolrRequest.METHOD.GET);
+  }
+
+  @Test
+  public void testAsyncPost() throws Exception {
+    super.testQueryAsync(SolrRequest.METHOD.GET);
+  }
+
+  @Test
+  public void testAsyncPut() throws Exception {
+    super.testQueryAsync(SolrRequest.METHOD.GET);
+  }
+
+  @Test
+  public void testAsyncException() throws Exception {
+    super.testAsyncExceptionBase();
+  }
+
+  @Test
   public void testFollowRedirect() throws Exception {
     final String clientUrl = getBaseUrl() + REDIRECT_SERVLET_PATH;
     try (Http2SolrClient client =
